# project-zer-supply-v2/site/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Zer Supply V2",
    "description": "Supply V2 is a dropship procurement platform that splits customer orders into vendor purchase orders, tracks shipments and receipts, resolves disputes, performs invoice matching, and protects every route with authentication, RBAC, and policy decisions.",
    "version": "2.0.0",
    "contact": {
      "name": "Project Zer Platform Team"
    }
  },
  "paths": {
...
    "/purchase-orders/{po_id}": {
      "get": {
        "tags": [
          "procurement"
        ],
        "summary": "Get Po",
        "operationId": "get_po_purchase_orders__po_id__get",
        "parameters": [
          {
            "name": "po_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Po Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tenant_demo",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user_demo",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "admin",
              "title": "X-Role"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/purchase-orders/{po_id}/acknowledge": {
      "post": {
        "tags": [
          "procurement"
        ],
        "summary": "Acknowledge Po",
        "operationId": "acknowledge_po_purchase_orders__po_id__acknowledge_post",
        "parameters": [
          {
            "name": "po_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Po Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tenant_demo",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user_demo",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "admin",
              "title": "X-Role"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AckDecisionIn"
                },
                "title": "Decisions"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/purchase-orders/{po_id}/vendor-disputes": {
      "post": {
        "tags": [
          "procurement"
        ],
        "summary": "Raise Vendor Dispute",
        "operationId": "raise_vendor_dispute_purchase_orders__po_id__vendor_disputes_post",
        "parameters": [
          {
            "name": "po_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Po Id"
            }
          },
          {
            "name": "x-vendor-access-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Vendor-Access-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorDisputeRaiseIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{order_id}/cancel-line": {
      "post": {
        "tags": [
          "procurement"
        ],
        "summary": "Cancel Order Line",
        "operationId": "cancel_order_line_orders__order_id__cancel_line_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tenant_demo",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user_demo",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "admin",
              "title": "X-Role"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelLineIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{order_id}/reallocate-line": {
      "post": {
        "tags": [
          "procurement"
        ],
        "summary": "Reallocate Order Line",
        "operationId": "reallocate_order_line_orders__order_id__reallocate_line_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tenant_demo",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user_demo",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "admin",
              "title": "X-Role"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReallocateLineIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AckDecisionIn": {
        "properties": {
          "po_line_id": {
            "type": "string",
            "title": "Po Line Id"
          },
          "accepted_quantity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Accepted Quantity"
          },
          "proposed_unit_price_minor": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Unit Price Minor"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "po_line_id",
          "accepted_quantity",
          "status"
        ],
        "title": "AckDecisionIn"
      },
      "CancelLineIn": {
        "properties": {
          "order_line_id": {
            "type": "string",
            "title": "Order Line Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "order_line_id",
          "reason"
        ],
        "title": "CancelLineIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReallocateLineIn": {
        "properties": {
          "order_line_id": {
            "type": "string",
            "title": "Order Line Id"
          },
          "new_vendor_id": {
            "type": "string",
            "title": "New Vendor Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "order_line_id",
          "new_vendor_id",
          "reason"
        ],
        "title": "ReallocateLineIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VendorDisputeRaiseIn": {
        "properties": {
          "po_line_id": {
            "type": "string",
            "title": "Po Line Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "accepted_quantity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Accepted Quantity"
          },
          "proposed_unit_price_minor": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Unit Price Minor"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "po_line_id",
          "status",
          "accepted_quantity",
          "reason"
        ],
        "title": "VendorDisputeRaiseIn"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "TenantHeaderAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-Id"
      },
      "InternalKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Internal-Api-Key"
      }
    }
  },
...
}